requirements.txt
Package                       Version
----------------------------- -------
asgiref                       3.8.1
dj-rest-auth                  6.0.0
Django                        4.2.11
django-cors-headers           4.3.1
django-rest-auth              0.9.5
djangorestframework           3.15.1
djangorestframework-simplejwt 5.3.1
pip                           24.0
psycopg                       3.1.18
psycopg2-binary               2.9.9
PyJWT                         2.8.0
setuptools                    58.0.4
six                           1.16.0
sqlparse                      0.5.0
typing_extensions             4.11.0
wheel                         0.43.0